services/pdf-extractor-python/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pdfplumber>=0.10.0
pdfminer.six>=20231228
spacy>=3.7.0
python-multipart>=0.0.6
pydantic>=2.5.0
es-core-news-md @ https://github.com/explosion/spacy-models/releases/download/es_core_news_md-3.7.0/es_core_news_md-3.7.0-py3-none-any.whl
